+++ Internal/Utility/Keyboard Elements/ProjectManager Up One Level.py
@@ -1,0 +1,6 @@
+#!/bin/env python
+# Set Basic Resolve Settings
+dvr = bmd.scriptapp("Resolve")
+fusion = dvr.Fusion()
+projectManager = dvr.GetProjectManager()
+projectManager.GotoParentFolder()